gcc 7.2 pedantic warning error fixes

+++ src/stream/ext_ogg.c
@@ -645,7 +645,7 @@ _aaxOggInitFormat(_driver_t *handle, unsigned char *oggbuf, size_t *bufsize)
 static int
 _getOggPageHeader(_driver_t *handle, uint32_t *header, size_t size)
 {
-   ssize_t bufsize = handle->oggBuffer->avail;
+   size_t bufsize = handle->oggBuffer->avail;
    uint32_t curr;
    int rv = 0;
 
@@ -836,7 +836,7 @@ _getOggPageHeader(_driver_t *handle, uint32_t *header, size_t size)
                rv += ch[27+i];
             }
 
-            if (rv <= bufsize)
+            if (rv <= (int)bufsize)
             {
                _aaxDataMove(handle->oggBuffer, NULL, rv);
                rv = 0;
